refactor(pi_validate): switch to subpackage; delete old flat files
Atomic switchover:
- Deleted yoyopod_cli/pi_validate.py (2940 LOC), _pi_validate_helpers.py
  (1350 LOC), and pi_validate_helpers.py (the public re-export shim).
- Renamed yoyopod_cli/_pi_validate_pkg -> yoyopod_cli/pi_validate.
- Replaced internal references to the temp package name with the final
  one (sed across the renamed tree).

main.py needs no edits — its `from yoyopod_cli import pi_validate as
_pi_validate` resolves to the new package. CLI --help surface diffs to
zero against pre-split baseline for root and all 8 subcommands. Gate
passes.

Tests (4 files) still need to migrate from attribute-access patterns
on the old flat module to the new submodule paths — covered by next
commits.
