+++ yoyopod_cli/pi_validate/_navigation_soak/__init__.py
@@ -1,21 +1,21 @@
 """Navigation soak utilities for pi_validate."""
 
-from yoyopod_cli._pi_validate_pkg._navigation_soak.handle import (
+from yoyopod_cli.pi_validate._navigation_soak.handle import (
     _NavigationSoakAppFactory,
     _NavigationSoakAppHandle,
     _YoyoPodAppNavigationSoakHandle,
     _default_app_factory,
 )
-from yoyopod_cli._pi_validate_pkg._navigation_soak.idle import (
+from yoyopod_cli.pi_validate._navigation_soak.idle import (
     run_navigation_idle_soak,
 )
-from yoyopod_cli._pi_validate_pkg._navigation_soak.plan import (
+from yoyopod_cli.pi_validate._navigation_soak.plan import (
     NavigationSoakError,
     NavigationSoakReport,
     NavigationSoakStep,
     build_navigation_soak_plan,
 )
-from yoyopod_cli._pi_validate_pkg._navigation_soak.runner import (
+from yoyopod_cli.pi_validate._navigation_soak.runner import (
     NavigationSoakFailure,
     NavigationSoakRunner,
     NavigationSoakStats,
