+++ yoyopod_cli/pi_validate/_navigation_soak/plan.py
