+++ yoyopod_cli/pi_validate/stability.py
@@ -1,4 +1,4 @@
-"""Lvgl validation subcommand."""
+"""Stability validation subcommand."""
 
 from __future__ import annotations
 
@@ -8,19 +8,16 @@ import typer
 
 from yoyopod_cli.common import configure_logging
 from yoyopod_cli.defaults import DEFAULT_TEST_MUSIC_TARGET_DIR
-from yoyopod_cli._pi_validate_pkg._navigation_soak import (
+from yoyopod_cli.pi_validate._navigation_soak import (
     NavigationSoakError,
     run_navigation_idle_soak,
 )
 
 
-def lvgl(
+def stability(
     config_dir: Annotated[
         str, typer.Option("--config-dir", help="Configuration directory to use.")
     ] = "config",
-    simulate: Annotated[
-        bool, typer.Option("--simulate", help="Run against simulation instead of hardware.")
-    ] = False,
     cycles: Annotated[
         int, typer.Option("--cycles", help="How many full transition cycles to run.")
     ] = 2,
@@ -36,14 +33,14 @@ def lvgl(
         bool,
         typer.Option(
             "--with-music",
-            help="Exercise playlist loading and now-playing actions during the soak.",
+            help="Also exercise playlist loading and now-playing actions during the soak.",
         ),
     ] = False,
     provision_test_music: Annotated[
         bool,
         typer.Option(
             "--provision-test-music/--no-provision-test-music",
-            help="Seed the deterministic validation music library before playback soak steps.",
+            help="Seed deterministic validation music before playback soak steps.",
         ),
     ] = True,
     test_music_dir: Annotated[
@@ -54,19 +51,17 @@ def lvgl(
         ),
     ] = "",
     skip_sleep: Annotated[
-        bool, typer.Option("--skip-sleep", help="Skip the sleep/wake exercise.")
+        bool, typer.Option("--skip-sleep", help="Skip the sleep and wake exercise.")
     ] = False,
     verbose: Annotated[bool, typer.Option("--verbose", help="Enable DEBUG logging.")] = False,
 ) -> None:
-    """Run a deterministic LVGL navigation and idle soak pass against YoYoPod."""
-    from loguru import logger
-
+    """Run a repeated navigation and idle stability pass on the target checkout."""
     configure_logging(verbose)
     resolved_music_dir = test_music_dir or DEFAULT_TEST_MUSIC_TARGET_DIR
     try:
         report = run_navigation_idle_soak(
             config_dir=config_dir,
-            simulate=simulate,
+            simulate=False,
             cycles=cycles,
             hold_seconds=hold_seconds,
             idle_seconds=idle_seconds,
@@ -76,6 +71,11 @@ def lvgl(
             test_music_dir=resolved_music_dir,
         )
     except NavigationSoakError as exc:
-        logger.error(f"LVGL soak failed: {exc}")
+        from loguru import logger
+
+        logger.error(f"Stability soak failed: {exc}")
         raise typer.Exit(code=1)
-    logger.info(f"LVGL soak passed: {report.summary()}")
+
+    from loguru import logger
+
+    logger.info(f"Stability soak passed: {report.summary()}")
